requirements.txt
# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0
ollama>=0.1.0

# Note: sqlite3 is part of Python's standard library (built-in)

# Optional: For future LangGraph integration
# langgraph>=0.2.0
# langchain>=0.3.0
# langchain-core>=0.3.0

